feat: rename directory repositorys to repositories
